feat: restructure backend
